+++ src/main/java/mpicbg/stitching/fusion/AveragePixelFusion.java
@@ -26,7 +26,7 @@ public class AveragePixelFusion implements PixelFusion
 	{ 
 		if ( count == 0 )
 			return 0;
-		return (float) (avg/(double)count); 
+		return (float) (avg/count); 
 	}
 
 	@Override
